requirements.txt
slack-sdk
google-generativeai
python-dotenv